test(popover): migrate position
